+++ src/preprocess/preprocess.hpp
@@ -79,7 +79,7 @@ namespace preprocess
     void resize_bilinear_gpu(float *d_tar, uint8_t *d_src, int tarW, int tarH, int srcH, int srcW, tactics tac, cudaStream_t stream);
     void init_process(int h, int w);
     void destroy_process();
-    __host__ __device__ void affine_transformation(float *trans_matrix, int src_x, int src_y, float *tar_x, float *tar_y);
+    __host__ __device__ void affine_transformation(float *trans_matrix, float src_x, float src_y, float *tar_x, float *tar_y);
 }; // namespace preprocess
 
 #endif //__PREPROCESS_HPP__
